Same changes as #1141 for parallel version.