@abhinavmehndiratta 2019-02-23
```julia
julia> nthreads()
2

julia> g = loadsnap(:ego_twitter_d)
{81306, 1768149} directed simple Int64 graph

julia> @benchmark pagerank(g)
BenchmarkTools.Trial: 
  memory estimate:  2.11 MiB
  allocs estimate:  20
  --------------
  minimum time:     49.038 ms (0.00% GC)
  median time:      49.944 ms (0.00% GC)
  mean time:        50.222 ms (0.12% GC)
  maximum time:     55.078 ms (0.78% GC)
  --------------
  samples:          100
  evals/sample:     1

julia> @benchmark old_pagerank(g)
BenchmarkTools.Trial: 
  memory estimate:  2.73 MiB
  allocs estimate:  35
  --------------
  minimum time:     61.254 ms (0.00% GC)
  median time:      62.067 ms (0.00% GC)
  mean time:        62.330 ms (0.11% GC)
  maximum time:     66.341 ms (0.71% GC)
  --------------
  samples:          81
  evals/sample:     1

```
Interestingly, the speedup here is not that much as in sequential probably because of resource sharing overhead.But it is still better than the previous version.
Btw the build is failing due to Label Propagation tests.