@abhinavmehndiratta 2019-02-25
> It says [here](https://codecov.io/gh/JuliaGraphs/LightGraphs.jl/commit/c1a125e01fb5688d85d6240b1b63503af3d9461b) the `let` statement is not being tests.

I think there is a problem with code coverage, since it is not possible that let block is not executed.