@simonschoelly 2019-03-15
The building passed now. In my eyes this looks good, what do you think @SohamTamba ? You are probably more familiar with the page rank algorithm than I am.