@SohamTamba 2019-03-18
Looks good
Can you post benchmarks (before/after) with the number of cores.

Also plz squash the commits into one.