@abhinavmehndiratta 2019-03-18
**Benchmarks**

NEW -
```julia
julia> nthreads()
2

julia> g = loadsnap(:ego_twitter_d)
{81306, 1768149} directed simple Int64 graph

julia> @benchmark Parallel.pagerank(g)
BenchmarkTools.Trial: 
  memory estimate:  2.73 MiB
  allocs estimate:  31
  --------------
  minimum time:     39.000 ms (0.00% GC)
  median time:      39.385 ms (0.00% GC)
  mean time:        39.548 ms (0.15% GC)
  maximum time:     41.672 ms (0.00% GC)
  --------------
  samples:          127
  evals/sample:     1

julia> g = loadsnap(:as_caida)
{26475, 53381} undirected simple Int64 graph

julia> @benchmark Parallel.pagerank(g)
BenchmarkTools.Trial: 
  memory estimate:  828.78 KiB
  allocs estimate:  28
  --------------
  minimum time:     14.183 ms (0.00% GC)
  median time:      14.557 ms (0.00% GC)
  mean time:        15.221 ms (1.79% GC)
  maximum time:     78.786 ms (81.26% GC)
  --------------
  samples:          329
  evals/sample:     1

julia> g = loadsnap(:facebook_combined)
{4039, 88234} undirected simple Int64 graph

julia> @benchmark Parallel.pagerank(g)
BenchmarkTools.Trial: 
  memory estimate:  127.30 KiB
  allocs estimate:  23
  --------------
  minimum time:     4.376 ms (0.00% GC)
  median time:      4.416 ms (0.00% GC)
  mean time:        4.518 ms (1.69% GC)
  maximum time:     67.222 ms (93.05% GC)
  --------------
  samples:          1105
  evals/sample:     1

```
OLD-
```julia
julia> nthreads()
2

julia> g = loadsnap(:ego_twitter_d)
{81306, 1768149} directed simple Int64 graph

julia> @benchmark Parallel.pagerank(g)
BenchmarkTools.Trial: 
  memory estimate:  2.73 MiB
  allocs estimate:  35
  --------------
  minimum time:     58.912 ms (0.00% GC)
  median time:      59.497 ms (0.00% GC)
  mean time:        59.944 ms (0.12% GC)
  maximum time:     65.630 ms (0.00% GC)
  --------------
  samples:          84
  evals/sample:     1

julia> g = loadsnap(:as_caida)
{26475, 53381} undirected simple Int64 graph

julia> @benchmark Parallel.pagerank(g)
BenchmarkTools.Trial: 
  memory estimate:  828.84 KiB
  allocs estimate:  31
  --------------
  minimum time:     20.783 ms (0.00% GC)
  median time:      21.653 ms (0.00% GC)
  mean time:        22.155 ms (1.76% GC)
  maximum time:     86.473 ms (74.56% GC)
  --------------
  samples:          226
  evals/sample:     1

julia> g = loadsnap(:facebook_combined)
{4039, 88234} undirected simple Int64 graph

julia> @benchmark Parallel.pagerank(g)
BenchmarkTools.Trial: 
  memory estimate:  127.36 KiB
  allocs estimate:  26
  --------------
  minimum time:     4.766 ms (0.00% GC)
  median time:      4.986 ms (0.00% GC)
  mean time:        5.121 ms (1.71% GC)
  maximum time:     69.064 ms (92.78% GC)
  --------------
  samples:          975
  evals/sample:     1
```