@simonschoelly 2019-03-22
Nice, that is quite some speedup.
@sbromberger I looks as if this can be merged.